% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eocusum_sim.R
\name{calceo}
\alias{calceo}
\title{Compute Expected minus Observed value}
\usage{
calceo(df, coeff, yemp = TRUE)
}
\arguments{
\item{df}{Data Frame. First column Parsonnet Score and second column outcome of each operation.}

\item{coeff}{Numeric Vector. Estimated coefficients \eqn{\alpha}{alpha} and \eqn{\beta}{beta}
from the binary logistic regression model.}

\item{yemp}{Logical. If \code{TRUE} use observed outcome value, if \code{FALSE} use estimated
binary logistc regression model.}
}
\value{
Returns a single value which is the difference between expected risk and observed
outcome.
}
\description{
Compute Expected minus Observed value.
}
\references{
Lovegrove J, Valencia O, Treasure T, Sherlaw-Johnson C and Gallivan S (1997).
 Monitoring the results of cardiac surgery by variable life-adjusted display.
 \emph{The Lancet}, \strong{350}(9085), pp. 1128–1130.

Poloniecki J, Valencia O and Littlejohns P (1998).
 Cumulative risk adjusted mortality chart for detecting changes in death rate: observational
 study of heart surgery. \emph{BMJ}, \strong{316}(7146), pp. 1697–1700.

Steiner S (2014). Risk-Adjusted Monitoring of Outcomes in Health Care.
 In Lawless JF (ed.), \emph{Statistics in Action}, pp. 225-242. Informa UK Limited.
}
\author{
Philipp Wittenberg
}
