% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{compute_vmask}
\alias{compute_vmask}
\title{Compute V-Masks arms, nose and alarm points}
\usage{
compute_vmask(z, d1, d2, theta1, theta2)
}
\arguments{
\item{z}{Numeric Vector. ...}

\item{d1}{Double. For the XYZ CUSUM Distance d from vertex of V-Mask. d=h/k}

\item{d2}{Double. For the XYZ CUSUM Distance d from vertex of V-Mask. d=h/k}

\item{theta1}{Double. Angle ...}

\item{theta2}{Double. Angle ...}
}
\value{
...
}
\description{
Function for plotting truncated symeterical/asymetrical vmask
}
\author{
Philipp Wittenberg
}
