% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eocusum_sim.R
\name{eocusum.arl.h.sim}
\alias{eocusum.arl.h.sim}
\title{Compute alarm threshold of EO-CUSUM control charts using simulation}
\usage{
eocusum.arl.h.sim(L0, df, k, coeff, m = 100, yemp = TRUE, side = 1,
  nc = 1, verbose = TRUE)
}
\arguments{
\item{L0}{double. In-Control ARL.}

\item{df}{DataFrame. First column are Parsonnet Score values within a range of zero to 100 representing
the preoperative patient risk. The second column are binary (0/1) outcome values of each operation.}

\item{k}{A double}

\item{coeff}{NumericVector. Estimated coefficients \eqn{\alpha}{alpha} and \eqn{\beta}{beta} from the binary
logistic regression model. For more information see details.}

\item{m}{integer. Number of simulation runs.}

\item{yemp}{boolean. Use emirical outcome value.}

\item{side}{character. Default is "low" to calculate ARL for the upper arm of the V-mask. If side = "up",
calculate the lower arm of the V-mask.}

\item{nc}{integer. Number of cores.}

\item{verbose}{boolean. If TRUE verbose output is included, if FALSE a quiet calculation of h is done.}
}
\value{
Returns a single value which is the distance d of the V-mask for a given ARL and \eqn{\theta}{theta}.
}
\description{
Compute alarm threshold of EO-CUSUM control charts using simulation.
}
\details{
The function \code{eocusum.arl.h.sim} determines the control limit for given in-control ARL (L0) by applying a
multi-stage search procedure which includes secant rule and the parallel version of \code{\link{eocusum.arl.sim}}
using \code{\link{mclapply}}.
}
\author{
Philipp Wittenberg
}
