% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eocusum_sim.R
\name{eocusum.arl.sim}
\alias{eocusum.arl.sim}
\title{Compute ARLs of EO-CUSUM control charts using simulation}
\usage{
eocusum.arl.sim(r, k, h, df, coeff, yemp = TRUE, side = "low")
}
\arguments{
\item{r}{int. Number of of simulation runs.}

\item{k}{double. Reference value of the CUSUM control chart.}

\item{h}{double. Decision interval (alarm limit, threshold) of the CUSUM control chart.}

\item{df}{DataFrame. First column Parsonnet Score and second column outcome of each operation.}

\item{coeff}{NumericVector. Estimated coefficients \eqn{\alpha}{alpha} and \eqn{\beta}{beta}
from the binary logistic regression model. For more information see details.}

\item{yemp}{boolean. If TRUE use observed outcome value, if FALSE use estimated binary logistc
regression model.}

\item{side}{character. Default is "low" to calculate ARL for the upper arm of the V-mask. If side = "up",
calculate the lower arm of the V-mask.}
}
\value{
Returns a single value which is the Run Length.
}
\description{
Compute ARLs of EO-CUSUM control charts using simulation.
}
\details{
Describe the algorithm for calulating in control run length ...
}
\references{
Steiner SH, Cook RJ, Farewell VT and Treasure T (2000).
 “Monitoring surgical performance using risk-adjusted cumulative sum charts.”
 \emph{Biostatistics}, \strong{1}(4), pp. 441-452.
 doi: \href{https://doi.org/10.1093/biostatistics/1.4.441}{10.1093/biostatistics/1.4.441}.
}
\author{
Philipp Wittenberg
}
