% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eocusum_sim.R
\name{eocusum.arloc.sim}
\alias{eocusum.arloc.sim}
\title{Compute Out of Control ARLs of EO-CUSUM control charts using simulation}
\usage{
eocusum.arloc.sim(r, k, h, df, coeff, coeff2, QS = 1, side = 1)
}
\arguments{
\item{r}{int. Number of of simulation runs.}

\item{k}{double. Reference value of the CUSUM control chart.}

\item{h}{double. Decision interval (alarm limit, threshold) of the CUSUM control chart.}

\item{df}{DataFrame. First column Parsonnet Score and second column outcome of each operation.}

\item{coeff}{NumericVector. Estimated coefficients \eqn{\alpha}{alpha} and \eqn{\beta}{beta}
from the binary logistic regression model. For more information see details.}

\item{coeff2}{NumericVector. Estimated coefficients \eqn{\alpha}{alpha} and \eqn{\beta}{beta}
from the binary logistic regression model of a resampled dataset. For more information see
details.}

\item{QS}{double. Defines the performance of a surgeon with the odds ratio ratio of death Q.}

\item{side}{int. If side = 1, calculate ARL for the upper arm of the V-mask. If side = 2,
calulate the lower arm of the V-mask.}
}
\value{
Returns a single value which is the Run Length.
}
\description{
Compute Out of Control ARLs of EO-CUSUM control charts using simulation.
}
\details{
Describe the algorithm for calulating in control run length ...
}
\examples{
\dontrun{
rm(list=ls())
require(VLAD2)
data("surgeons"); data("s5000")
df1 <- subset(surgeons, select=c(parsonnet, record1))
df2 <- subset(s5000, select=c(parsonnet, record1))
## estimate coefficients from logit model
coeff1 <- round(coef(glm(record1~parsonnet, data=df1, family="binomial")), 3)
coeff2 <- round(coef(glm(record1~parsonnet, data=df2, family="binomial")), 3)

## Serial simulation
## set seed for reproducibility
RNGkind("L'Ecuyer-CMRG") # RNGkind("Mersenne-Twister")
set.seed(1234)
m <- 10^3
RLS <- do.call(c, lapply(1:m, eocusum.arloc.sim, h=4.498, k=0, df=df1, side=1, coeff=coeff1,
                         coeff2=coeff2))
data.frame(cbind("ARL"=mean(RLS), "ARLSE"=sd(RLS)/sqrt(m)))
## ARL=366.697; ARLSE=9.457748
## Parallel simulation (FORK)
## set seed for reproducibility
RNGkind("L'Ecuyer-CMRG")
set.seed(1234); parallel::mc.reset.stream()
m <- 10^3
RLS <- simplify2array(parallel::mclapply(1:m, eocusum.arloc.sim, h=4.498, k=0, df=df1, side=1,
                                         coeff=coeff1, coeff2=coeff2,
                                         mc.cores=parallel::detectCores()))
data.frame(cbind("ARL"=mean(RLS), "ARLSE"=sd(RLS)/sqrt(m)))
## ARL=372.133; ARLSE=9.967791

## Parallel simulation (PSOCK)
## set seed for reproducibility
RNGkind("L'Ecuyer-CMRG")
no_cores <- parallel::detectCores()
cl <- parallel::makeCluster(no_cores)
parallel::clusterSetRNGStream(cl, iseed = 1234)
side <- 1
h_vec <- 4.498
QS_vec <- 1
m <- 10^3
k <- 0
parallel::clusterExport(cl, c("h_vec", "eocusum.arloc.sim", "df1", "coeff1", "coeff2",
                              "QS_vec", "side", "k"))
time <- system.time( {
RLS <- array(NA, dim=c( length(QS_vec), length(h_vec), m))
for (h in h_vec) {
  for (QS in QS_vec) {
    cat(h, " ", QS, "\\n")
    RLS[which(QS_vec==QS), which(h==h_vec), ] <- parallel::parSapply(cl, 1:m, eocusum.arloc.sim,
                                                                     side=side, QS=QS, h=h, k=k,
                                                                     df=df1,  coeff=coeff1,
                                                                     coeff2=coeff2,
                                                                      USE.NAMES=FALSE)
    }
  }
} )
ARL <- apply(RLS, c(1, 2), mean)
ARLSE <- sqrt(apply(RLS, c(1, 2), var)/m)
print(list(ARL, ARLSE, time))
parallel::stopCluster(cl)
## ARL=371.361; ARLSE=9.899058
}
}
\references{
Steiner SH, Cook RJ, Farewell VT and Treasure T (2000).
 “Monitoring surgical performance using risk-adjusted cumulative sum charts.”
 \emph{Biostatistics}, \strong{1}(4), pp. 441-452.
 doi: \href{https://doi.org/10.1093/biostatistics/1.4.441}{10.1093/biostatistics/1.4.441}.
}
\author{
Philipp Wittenberg
}
