% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eocusum_sim.R
\name{eocusum_crit_sim}
\alias{eocusum_crit_sim}
\title{Compute alarm threshold of EO-CUSUM control charts using simulation}
\usage{
eocusum_crit_sim(L0, pmix, k, RQ = 1, side = "low", yemp = FALSE,
  m = 10000, nc = 1, hmax = 30, jmax = 4, verbose = FALSE)
}
\arguments{
\item{L0}{Double. Prespecified in-control Average Run Length.}

\item{pmix}{Data Frame. A three column data frame. First column is the operation outcome.
Second column are the predicted probabilities from the risk model. Third
 column can be either the predicted probabilities from the risk model or TODO .............}

\item{k}{Double. Reference value of the CUSUM control chart. Either \code{0} or a positive
value. Can be determined with function \code{\link{optimal_k}}.}

\item{side}{Character. Default is \code{"low"} to calculate ARL for the upper arm of the V-mask.
If side = \code{"up"}, calculate the lower arm of the V-mask.}

\item{yemp}{Logical. If \code{TRUE}, use emirical outcome values, else use model.}

\item{m}{Integer. Number of simulation runs.}

\item{nc}{Integer. Number of cores used for parallel processing.}

\item{hmax}{Integer. Maximum value of \code{h} for the grid search.}

\item{jmax}{Integer. Number of digits for grid search.}

\item{verbose}{Logical. If \code{TRUE} verbose output is included, if \code{FALSE} a quiet
calculation of \code{h} is done.}
}
\value{
Returns a single value which is the control limit \code{h} for a given in-control ARL.
}
\description{
Compute alarm threshold of EO-CUSUM control charts using simulation.
}
\details{
Determines the control limit ("\code{h}") for given in-control ARL (\code{"L0"})
applying a grid search using \code{\link{eocusum_arl_sim}} and \code{\link{parSapply}}.
}
\author{
Philipp Wittenberg
}
