% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racusum_beta_mc.R
\name{FWT2}
\alias{FWT2}
\title{scascas}
\usage{
FWT2(w, RA, g0, g1, shape1, shape2, RQ)
}
\arguments{
\item{w}{...}

\item{RA}{Double. Odds ratio of death under the alternative hypotheses. Detecting deterioration
in performance with increased mortality risk by doubling the odds Ratio \code{RA = 2}. Detecting
 improvement in performance with decreased mortality risk by halving the odds ratio of death
 \code{RA = 1/2}. Odds ratio of death under the null hypotheses is \code{1}.}

\item{g0}{...}

\item{g1}{...}

\item{shape1}{Double. Shape parameter \eqn{\alpha}{alpha} \code{> 0} of the beta distribution.}

\item{shape2}{Double. Shape parameter \eqn{\beta}{beta} \code{> 0} of the beta distribution.}

\item{RQ}{...}
}
\description{
Compute the log-leikelie
}
\author{
Philipp Wittenberg
}
