% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GandyKvaloy_P1}
\alias{GandyKvaloy_P1}
\title{cardiac-surgery dataset from spcadjust R-package.}
\format{A data frame with .... rows and ... variables:}
\usage{
data(GandyKvaloy_P1)
}
\description{
......adcmkdvsdvdv .....
}
\details{
\itemize{
  \item{S1}{... TODO}
  \item{N1}{... TODO}
  \item{X1}{... TODO}
}
}
\keyword{datasets}
