% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racusum_beta_int.R
\name{racusum_beta_arl_int}
\alias{racusum_beta_arl_int}
\title{Compute ARL of RA-CUSUM control charts assuming a beta distributed patient mix with using
 collocation methods.}
\usage{
racusum_beta_arl_int(h, N, RA, RQ, g0, g1, shape1, shape2, pw)
}
\arguments{
\item{h}{Double. \code{h} is the control limit (>\code{0}).}

\item{N}{Integer. Number of quadrature nodes, dimension of the resulting linear equation system
is equal TODO ???.}

\item{RA}{...}

\item{RQ}{...}

\item{g0}{Double. ...}

\item{g1}{Double. ...}

\item{shape1}{Double. Shape parameter \eqn{\alpha}{alpha} \code{> 0} of the beta distribution.}

\item{shape2}{Double. Shape parameter \eqn{\beta}{beta} \code{> 0} of the beta distribution.}

\item{pw}{Logical. Pice-wise collocation yes or no.}
}
\value{
Returns a single value which is the Run Length.
}
\description{
Compute ARL of RA-CUSUM control charts assuming a beta distributed patient mix with using
 collocation methods.
}
\examples{
\dontrun{
library(vlad)
racusum_beta_arl_int(h=4.5, RA=2, RQ=1, shape1=1, shape2=6, g0=-3.6798, g1=0.0768*71, N=50,
 pw=FALSE)
}
}
\author{
Philipp Wittenberg
}
