% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racusum_discretebeta_sim.R
\name{racusum_discretebeta_crit_sim}
\alias{racusum_discretebeta_crit_sim}
\title{Compute alarm threshold of RA-CUSUM control charts using simulation}
\usage{
racusum_discretebeta_crit_sim(L0, shape1, shape2, coeff, rs = 72, RA = 2,
  RQ = 1, nc = 1, hmax = 30, jmax = 4, m = 10000, verbose = FALSE)
}
\arguments{
\item{L0}{Double. Prespecified in-control Average Run Length.}

\item{shape1}{Double. Shape parameter \eqn{\alpha}{alpha} \code{> 0} of the beta distribution.}

\item{shape2}{Double. Shape parameter \eqn{\beta}{beta} \code{> 0} of the beta distribution.}

\item{coeff}{Numeric Vector. Estimated intercept and slope coefficients
from a binary logistic regression model.}

\item{rs}{Integer. Number of intervals between \code{0} and the maximum risk score.}

\item{RA}{Double. Odds ratio of death under the alternative hypotheses. Detecting deterioration
in performance with increased mortality risk by doubling the odds Ratio \code{RA = 2}. Detecting
 improvement in performance with decreased mortality risk by halving the odds ratio of death
 \code{RA = 1/2}.}

\item{RQ}{Double. Defines the performance of a surgeon with the odds ratio ratio of death.
\code{Q}.}

\item{nc}{Integer. Number of cores used for parallel processing. Value is passed to
\code{\link{parSapply}}.}

\item{hmax}{Integer. Maximum value of \code{h} for the grid search.}

\item{jmax}{Integer. Number of digits for grid search.}

\item{m}{Integer. Number of simulation runs.}

\item{verbose}{Logical. If \code{TRUE} verbose output is included, if \code{FALSE} a quiet
calculation of \code{h} is done.}
}
\value{
Returns a single value which is the control limit \code{h} for a given in-control ARL.
}
\description{
Compute alarm threshold of risk-adjusted cumulative sum control charts using
 simulation.
}
\details{
Determines the control limit ("\code{h}") for given in-control ARL (\code{"L0"})
applying a grid search using \code{\link{racusum_discretebeta_arl_sim}} and
\code{\link{parSapply}}.
}
\examples{
\dontrun{
library(vlad)
racusum_discretebeta_crit_sim(L0=7500, shape1=.61, shape2=4.09, rs=(71+1),
coeff=c(-3.6798, .0768), RA=2, RQ=1, nc=4, verbose=TRUE, m=1e3)
}

}
\author{
Philipp Wittenberg
}
