% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rockettails}
\alias{rockettails}
\title{Rockettails for VLAD}
\usage{
rockettails(alpha, s, coeff)
}
\arguments{
\item{alpha}{double.}

\item{s}{double. Parsonnet score.}

\item{coeff}{NumericVector. Estimated coefficients \eqn{\alpha}{alpha} and \eqn{\beta}{beta}
from the binary logistic regression model. For more information see details.}
}
\value{
....
}
\description{
Rockettails for VLAD.
}
\examples{
\dontrun{
library("dplyr")
library("tidyr")
library("ggplot2")
data("cardiacsurgery", package = "spcadjust")
SALL <- cardiacsurgery \%>\% rename(s = Parsonnet) \%>\%
  mutate(y = ifelse(status == 1 & time <= 30, 1, 0),
         phase = factor(ifelse(date < 2*365, "I", "II")))

SALLI <- filter(SALL, phase == "I")
coeff1 <- round(coef(glm(y ~ s, data = SALLI, family = "binomial")), 3)

S2 <- filter(SALL, surgeon == 2)
S2I <- filter(S2, phase == "I") \%>\% select(s, y)
S2II <- filter(S2, phase == "II") \%>\% select(s, y)

rt <- rockettails(alpha = 0.05, s = S2II[, "s"] , coeff = coeff1)
EO <- sapply(1:nrow(S2II), function(i) calceo(df = S2II[i, c("s", "y")], coeff = coeff1,
             yemp = TRUE))
df1 <- data.frame(cbind("n" = 1:nrow(S2II), "cEO" = cumsum(EO), rt)) \%>\%
  gather(key = variable, value = value, c(-n))

ggplot(df1, aes(x = n, y = value, group = variable)) +
  geom_line() +
  geom_point(data = filter(df1, variable == "cEO"), col = "red") +
  geom_hline(yintercept = 0, linetype = "dashed", col = "darkgreen") +
  theme_bw() + labs(x = "Patient number n", y = "CUSUM E-O") +
  geom_hline(yintercept=0, linetype="dashed", col="darkgreen") +
  scale_y_continuous(sec.axis=dup_axis(name=NULL, labels=NULL)) +
  scale_x_continuous(sec.axis=dup_axis(name=NULL, labels=NULL))
}
}
\author{
Philipp Wittenberg
}
\keyword{keyword1}
\keyword{keyword2}
