% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{surgery}
\alias{surgery}
\title{Surgical outcome data.}
\format{
A data frame with 2500 rows and 2 variables:
\describe{
  \item{s}{Risk scores}
  \item{y}{Binary operation outcome (0=survival, 1=death)}
}
}
\usage{
surgery
}
\description{
A data set with the risk scores and surgical outcomes of 2,500 patients.
}
\keyword{datasets}
