% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{trafo}
\alias{trafo}
\title{Box-Cox transformation of data.}
\usage{
trafo(delta, x)
}
\arguments{
\item{delta}{Numeric. Box-Cox transformation parameter.}

\item{x}{Numceric Vector. Parsonnet Score values within a range of 0 to 100 representing the
preoperative patient risk.}
}
\value{
Returns a transformed Numeric vector.
}
\description{
Box-Cox transformation of data.
}
\author{
Philipp Wittenberg
}
