% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{vlad-package}
\alias{vlad-package}
\title{Variable Life Adjusted Display}
\description{
Contains functions to set up risk-adjusted quality control charts in health care. For the
variable life adjusted display (VLAD) and RA-CUSUM based on log-likelihood ratio statistic parameters
such as the Average Run Length (ARL) or control limits can be computed.
}
\author{
Philipp Wittenberg <pwitten@hsu-hh.de>
}
