% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{VMASK3}
\alias{VMASK3}
\title{Vmask3}
\usage{
VMASK3(A, B, d1, d2, theta1, theta2, Sn, seg)
}
\arguments{
\item{A}{...}

\item{B}{...}

\item{d1}{Double. For the XYZ CUSUM Distance d from vertex of V-Mask. d=h/k}

\item{d2}{Double. For the XYZ CUSUM Distance d from vertex of V-Mask. d=h/k}

\item{theta1}{Double. Angle ...}

\item{theta2}{Double. Angle ...}

\item{Sn}{...}

\item{seg}{Logical. ...}
}
\value{
...
}
\description{
Helper function to compute truncated symeterical/asymetrical vmask
}
\author{
Philipp Wittenberg
}
