% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racusum_sim.R
\name{cusum_arl_sim}
\alias{cusum_arl_sim}
\title{Compute ARLs of Bernoulli CUSUM control charts using simulation}
\usage{
cusum_arl_sim(r, h, df, R0 = 1, RA = 2)
}
\arguments{
\item{r}{Integer Vector. Number of runs.}

\item{h}{Double. Control Chart limit for detecting deterioration/improvement.}

\item{df}{Data Frame. First column are Parsonnet Score values within a range of \code{0} to
\code{100} representing the preoperative patient risk. The second column are binary (\code{0/1})
 outcome values of each operation.}

\item{R0}{Double. Odds ratio of death under the null hypotheses.}

\item{RA}{Double. Odds ratio of death under the alternative hypotheses.}
}
\description{
Compute ARLs of Bernoulli cumulative sum control charts using simulation.
}
\examples{
\donttest{

# This function is deprecated. See bcusum_arl_sim() instead.

 }
}
\author{
Philipp Wittenberg
}
