% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racusum_mc.R
\name{racusum_arl_mc}
\alias{racusum_arl_mc}
\title{Compute ARLs of RA-CUSUM control charts using Markov chain approximation}
\usage{
racusum_arl_mc(pmix, RA, RQ, h, scaling = 600, rounding = TRUE,
  method = "Toep")
}
\arguments{
\item{pmix}{Numeric Matrix. Patient mix defined}

\item{RA}{Double. Odds ratio of death under the alternative hypotheses. Detecting deterioration
in performance with increased mortality risk by doubling the odds Ratio \code{RA = 2}. Detecting
 improvement in performance with decreased mortality risk by halving the odds ratio of death
 \code{RA = 1/2}. Odds ratio of death under the null hypotheses is \code{1}.}

\item{RQ}{Double. Defines the true performance of a surgeon with the odds ratio ratio of death
\code{RQ}. Use \code{RQ = 1} to compute the in-control ARL and other values to compute the
out-of-control ARL.}

\item{h}{Double. \code{UCL} is the control limit (>\code{0}).}

\item{scaling}{Double. The \code{scaling} parameter controls the quality of the approximation,
larger values achieve higher accuracy but increase the computation burden (larger transition
probability matrix).}

\item{rounding}{Logical. If \code{TRUE} a paired rounding implementation similar to
\emph{Webster and Pettitt (2007)} is used, if \code{FALSE} a simple rounding method of
\emph{Steiner et al. (2000)} is used.}

\item{method}{Character. If \code{method = "Toep"} a combination of Sequential Probability Ratio
Test and Toeplitz matrix structure is used to calculate the ARL. \code{"ToepInv"} computes the
inverted matrix using Toeplitz matrix structure. \code{"BE"} solves a linear equation system
using classical approach of \emph{Brook and Evans (1972)} to calculate the ARL.}
}
\value{
Returns a single value which is the Run Length.
}
\description{
Compute ARLs of RA-CUSUM control charts using Markov chain approximation.
}
\author{
Philipp Wittenberg
}
