% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racusum_mc.R
\name{racusum_crit_mc}
\alias{racusum_crit_mc}
\title{Compute alarm threshold of RA-CUSUM control charts using Markov chain approximation}
\usage{
racusum_crit_mc(pmix, L0, RA, RQ, scaling = 600, rounding = TRUE,
  method = "Toep", verbose = FALSE)
}
\arguments{
\item{pmix}{Data Frame. TODO}

\item{L0}{Double. Prespecified Average Run Length.}

\item{RA}{Double. Odds ratio of death under the alternative hypotheses. Detecting deterioration
in performance with increased mortality risk by doubling the odds Ratio \code{RA = 2}. Detecting
 improvement in performance with decreased mortality risk by halving the odds ratio of death
 \code{RA = 1/2}. Odds ratio of death under the null hypotheses is \code{1}.}

\item{RQ}{Double. Defines the true performance of a surgeon with the odds ratio ratio of death
\code{RQ}. Use \code{RQ = 1} to compute the in-control ARL and other values to compute the
out-of-control ARL.}

\item{scaling}{Double. The \code{scaling} parameter controls the quality of the approximation,
larger values achieve higher accuracy but increase the computation burden (larger transition
probability matrix).}

\item{rounding}{Logical. If \code{TRUE} a paired rounding implementation similar to
\emph{Webster and Pettitt (2007)} is used, if \code{FALSE} a simple rounding method of
\emph{Steiner et al. (2000)} is used.}

\item{method}{Character. If \code{method = "Toep"} a combination of Sequential Probability Ratio
Test and Toeplitz matrix structure is used to calculate the ARL. \code{"ToepInv"} computes the
inverted matrix using Toeplitz matrix structure. \code{"BE"} solves a linear equation system
using classical approach of \emph{Brook and Evans (1972)} to calculate the ARL.}

\item{verbose}{Logical. If \code{FALSE} a quiet calculation of \code{h} is done. If \code{TRUE}
verbose output of the search procedure (see details) is included.}
}
\value{
Returns a single value which is the control limit \code{h} for a given In-control ARL.
}
\description{
Compute alarm threshold of RA-CUSUM control charts using Markov chain
approximation.
}
\details{
Determines the control limit for given in-control ARL (\code{"L0"}) using
\code{\link{racusum_arl_mc}} by applying a multi-stage search procedure with the following
steps:
\describe{
  \item{(i)}{Calculate ARL for an initial \code{h = 1}.}
  \item{(ii)}{Calculate ARL by increasing \code{h} by 1 as long as resulting ARL is smaller than
   \code{L0} and \code{h} smaller than \code{6}.}
  \item{(iii)}{If, for \code{h = 6} the resulting ARL is still smaller than \code{L0}, use
  linear interpolation the get a new pair of \code{h} and Arl.}
  \item{(iv)a}{Calculate ARL by increasing \code{h} by \code{1} as long as resulting ARL is
  smaller than \code{L0}.}
  \item{(iv)b}{If, ARL if ARL is greater than ARL (iii), calculate ARL by decreasing \code{h} by
   \code{1} as long as resulting ARL is greater or equal than \code{L0}.}
  \item{(v)}{Use scant method with absolue error greater than 1e-4 and \code{h} error greater
  than \code{1e-6}.}
  \item{(vi)}{TODO ...}
  \item{(vii)}{TODO ...}
}
}
\author{
Philipp Wittenberg
}
