% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racusum_sim.R
\name{racusum_crit_sim}
\alias{racusum_crit_sim}
\title{Compute alarm threshold of RA-CUSUM control charts using simulation}
\usage{
racusum_crit_sim(L0, df, coeff, R0 = 1, RA = 2, m = 100,
  yemp = TRUE, nc = 1, jmax = 4, verbose = FALSE)
}
\arguments{
\item{L0}{Double. Prespecified in-control Average Run Length.}

\item{df}{Data Frame. First column are Parsonnet Score values within a range of \code{0} to
\code{100} representing the preoperative patient risk. The second column are binary (\code{0/1})
 outcome values of each operation.}

\item{coeff}{Numeric Vector. Estimated coefficients \eqn{\alpha}{alpha} and \eqn{\beta}{beta}
from the binary logistic regression model.}

\item{R0}{Double. Odds ratio of death under the null hypotheses.}

\item{RA}{Double. Odds ratio of death under the alternative hypotheses. Detecting deterioration
in performance with increased mortality risk by doubling the odds Ratio \code{RA = 2}.
Detecting improvement in performance with decreased mortality risk by halving the odds ratio of
 death \code{RA = 1/2}.}

\item{m}{Integer. Number of simulation runs.}

\item{yemp}{Logical. If \code{TRUE}, use emirical outcome values, else use model.}

\item{nc}{Integer. Number of cores used for parallel processing.}

\item{jmax}{Integer. Number of digits for grid search.}

\item{verbose}{Logical. If \code{TRUE} verbose output is included, if \code{FALSE} a quiet
calculation of \code{h} is done.}
}
\value{
Returns a single value which is the control limit \code{h} for a given in-control ARL.
}
\description{
Compute alarm threshold of risk-adjusted cumulative sum control charts using
 simulation.
}
\details{
The function \code{racusum_crit_sim} determines the control limit \code{h} for given
 in-control ARL (\code{L0}) by applying a multi-stage search procedure which includes secant
 rule and the parallel version of \code{\link{racusum_arl_sim}} using \code{\link{mclapply}}.
}
\examples{
\dontrun{
library(vlad)
library("dplyr")
data("cardiacsurgery", package = "spcadjust")

## preprocess data to 30 day mortality and subset phase I (In-control) of surgeons 2
S2I <- cardiacsurgery \%>\% rename(s = Parsonnet) \%>\%
  mutate(y = ifelse(status == 1 & time <= 30, 1, 0),
         phase = factor(ifelse(date < 2*365, "I", "II"))) \%>\%
  filter(phase == "I", surgeon == 2) \%>\% select(s, y)

## estimate coefficients from logit model
coeff1 <- round(coef(glm(y ~ s, data = S2I, family = "binomial")), 3)

## control limit for detecting deterioration RA = 2:
racusum_crit_sim(L0 = 740, df = S2I, coeff = coeff1, m = 10^3, nc = 4)
}
}
\references{
Steiner SH, Cook RJ, Farewell VT and Treasure T (2000).
 Monitoring surgical performance using risk-adjusted cumulative sum charts.
 \emph{Biostatistics}, \strong{1}(4), pp. 441--452.

Wittenberg P, Gan FF, Knoth S (2018).
A simple signaling rule for variable life-adjusted display derived from
an equivalent risk-adjusted CUSUM chart.
\emph{Statistics in Medicine}, \strong{37}(16), pp 2455--2473.
}
\author{
Philipp Wittenberg
}
