% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{vlad-package}
\alias{vlad-package}
\title{Variable Life Adjusted Display}
\description{
Contains functions to set up risk-adjusted quality control charts
in health care. For the variable life adjusted display (VLAD) proposed by
Lovegrove et al. (1997) <doi:10.1016/S0140-6736(97)06507-0> and the
risk-adjusted cumulative sum chart based on log-likelihood ratio statistic
introduced by Steiner et al. (2000) <doi:10.1093/biostatistics/1.4.441> the
average run length and control limits can be computed.
}
\author{
Philipp Wittenberg and Sven Knoth
}
